refactor: remove dead code found by fallow
- Delete unused barrel files (client/index.ts, format/index.ts, config/index.ts, utils/url.ts)
- Un-export 7 functions used only internally (pageUrl, EXPORT_MARKER, writeExportMarker, isExportDirectory, exportRecursive, printTreeNodes, resolveProfile)
- Delete 3 unused type interfaces (CreateBlogPostRequest, UpdateBlogPostRequest, CopyTreeResult)
- Remove 6 unused dependencies (form-data, html-to-text, markdown-it, ora, axios-mock-adapter, @types/html-to-text)
- Update imports from barrel paths to direct module paths
- Add .fallowrc.json with FP suppressions for test files and dynamically dispatched class members
- Add .gitignore entry for .fallow cache

Fallow results: 59 → 23 issues (remaining are 1 test-only export + 22 dynamic dispatch FPs)
All 83 tests pass, lint clean, typecheck clean.
