+++ .gitignore
@@ -86,3 +86,4 @@ Thumbs.db
 
 # Configuration files with sensitive data
 config.json
+.fallow/
