+++ package.json
@@ -37,19 +37,13 @@
     "axios": "^1.15.0",
     "chalk": "^4.1.2",
     "commander": "^11.1.0",
-    "form-data": "^4.0.5",
-    "html-to-text": "^9.0.5",
-    "inquirer": "^8.2.6",
-    "markdown-it": "^14.1.0",
-    "ora": "^5.4.1"
+    "inquirer": "^8.2.6"
   },
   "devDependencies": {
     "@biomejs/biome": "^2.4.13",
     "@types/commander": "^2.12.5",
-    "@types/html-to-text": "^9.0.4",
     "@types/inquirer": "^9.0.9",
     "@types/node": "^25.6.0",
-    "axios-mock-adapter": "^2.1.0",
     "tsup": "^8.5.1",
     "typescript": "^6.0.3"
   },
