--- src/client/index.ts
@@ -1,17 +1,0 @@
-export type { AttachmentsClient } from './attachments.js'
-export { DefaultAttachmentsClient } from './attachments.js'
-export { BlogClient } from './blog.js'
-export type { CommentsClient } from './comments.js'
-export { DefaultCommentsClient } from './comments.js'
-export { HttpClient } from './http.js'
-export type { LabelsClient } from './labels.js'
-export { DefaultLabelsClient } from './labels.js'
-export type { PagesClient } from './pages.js'
-export { DefaultPagesClient } from './pages.js'
-export type { PropertiesClient } from './properties.js'
-export { DefaultPropertiesClient } from './properties.js'
-export type { SearchClient, SearchOptions } from './search.js'
-export { DefaultSearchClient } from './search.js'
-export type { SpacesClient } from './spaces.js'
-export { DefaultSpacesClient } from './spaces.js'
-export type * from './types.js'
