--- src/config/index.ts
@@ -1,21 +1,0 @@
-export {
-  CONFIG_DIR,
-  CONFIG_FILE,
-  fetchCloudId,
-  getConfig,
-  loadConfig,
-  normalizeApiPath,
-  normalizeProtocol,
-} from './loader'
-export { deleteProfile, initConfig, isValidProfileName, listProfiles, setActiveProfile } from './profiles'
-export type {
-  AppConfig,
-  AuthConfig,
-  BasicAuthConfig,
-  BearerAuthConfig,
-  CookieAuthConfig,
-  MtlsAuthConfig,
-  ProfileConfig,
-  ResolvedConfig,
-  ServiceAccountAuthConfig,
-} from './types'
