--- src/format/index.ts
@@ -1,7 +1,0 @@
-export {
-  formatJson,
-  formatPageInfo,
-  formatProfiles,
-  formatSearchResults,
-  formatSpaces,
-} from './output'
